TST: move Parameters tests to dedicated module
