--- tests/test_errors.py
@@ -1,10 +1,0 @@
-import pytest
-
-from nonos.api import Parameters
-
-
-def test_init_params_wo_a_file():
-    with pytest.raises(
-        FileNotFoundError, match=r"idefix.ini, pluto.ini, variables.par not found"
-    ):
-        Parameters()
